Introduce set_error() function and fix error handling/formatting
Hi,
this is the patch to switch error reporting facility to set_error()
function which is having the original code. This is useful to set
error from every place we need instead of setting it just by libvirt
error handler. This was mainly done to bail with error in the case
you call libvirt function that's not supported by your version of
libvirt (using LIBVIR_VERSION_NUMBER constant).

Signed-off-by: Michal Novotny <minovotn@redhat.com>