Introduce libvirt_get_iso_images() API and extend PHPInfo() output
This is the patch to introduce the libvirt_get_iso_images() to look
for the ISO images on the path specified as the first argument or
the path in the configuration (defaults to /var/lib/libvirt/images/iso).
This patch also extends the PHPInfo() output to show the ISO Image path
status, conversion from long long to string status and also features
supported on the system, like ability to get screenshots using
external utilities and similar.

Signed-off-by: Michal Novotny <minovotn@redhat.com>