Implement debug support using libvirt_logfile_set() API function
This is a very useful patch to implement the debug logging for libvirt-php
module. The debug logging can be turned on using the libvirt_logfile_set()
API function that accepts both filename and maximum log file size in KiB.
Debug support have to be compiled into the module which is enabled by in
this version with default set to logging disabled.

Also, version number in the PHPInfo() has been changed from 0.4.1 to
0.4.2.

Signed-off-by: Michal Novotny <minovotn@redhat.com>