configure: Prefer /bin/echo before printf
For some reason, configure think it's good to use 'printf %s\n'
as $(ECHO). Well, it's not. Problem is, the \n at the end - it
does not do what you think. It merely adds 'n' at the end of each
printed string. Therefore, use echo which does not need anything
special.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>