libvirt_network_get_information: Be more tolerant to missing info
This function is really build in a way that only NATed networks
can pass the function. Any other networks (and there's plenty of
them) will make it fail immediately. So while we are in PHP, we
can utilize array_key_exist() function and thus just fill just
those info that we know for now.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>