Use zend_long instead of zend_ulong64
In PHP 7.0, the zend_(u)long64 definition was removed in favor
of zend_(u)long changing depending on whether PHP was compiled on
a 64-bit architecture or not.

The patch set that originally ported libvirt-php to PHP7+ did not
use zend_ulong64. However, it was altered to use it during the review,
which broke PHP 7.0 support. This fixes that by using zend_long and
checking for if the value is invalid and returning false when that happens.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>