examples: Properly detect if we're connected
So we have this special class Libvirt implemented in
examples/libvirt.php. It's aim is to wrap some low-level libvirt
APIs into slightly more advanced methods. Now, when instantiating
the class, connection URI can be passed in which case the class
will connect to it right in the constructor. However, later in
the code, the returned value (which is an object reference) is
compared against false, which will never ever be true.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>