Don't export any symbols but get_module
This function is the entry point to the whole module. Anything
else will be registered in it and therefore it is okay if the
rest of symbols are local. Moreover, we are exporting our own
internal functions which may clash with other libraries suffering
from the same problem. For instance, there has been a bug report
on the list recently [1] where libwsman_client is requiring
generate_uuid symbol (from its another library lbwsman) but what
a coincidence - libvirt-php has such function too (which is
purely internal) and worse - we export it.

1: https://www.redhat.com/archives/libvir-list/2016-September/msg00053.html

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>