examples: Sort domains
In the examples/index.php we list all domains/networks/.. defined
for the connection we have. However, the objects are printed out
in the order as provided by libvirt, so basically a random order.
It may be more user friendly if we sort the array of objects
before printing it out.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>