libvirt-php.c: Fix coding style issues: if then else statement style issues
Follow the following rules in if - then - else statements:

1) The opening brackets stays in the same line with the if statement

2) The closing brace is empty on a line of its own except in
the cases where it is followed by a continuation of the same statement.
Example:
if (x == y) {
    ...
} else {
    ...
}

3) Do not unnecessarily use braces where a single statement will do.
However, if one branch of a conditional statement is a single
statement and the other multiline use braces in both branches.
example:

if (condition) {
	do_this();
	do_that();
} else {
	otherwise();
}

Signed-off-by: Katerina Koukiou <k.koukiou@gmail.com>