libvirt_domain_get_screenshot_api: Don't unlink screenshot file on success
In 5120ba28a2021 I've tried to fix couple of problems with this
API (all kinds of resource leaks, bad conditions to some checks
and so on). However, my clumsy fix introduced unlink even to a
success path so after all callers were left with a non-existent
path to screenshot file.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>