libvirt-php: Introduce get_node_string_from_xpath()
So far we have get_string_from_xpath() which works in most cases,
but it is no good for the next commit. Problem with
get_string_from_xpath() is that it immediately converts the
result using xmlNodeListGetString() which is like wrapping xpath
into string() like this: "string(xpath)". While this works in
most cases when we query for just an attribute value it fails
when we want to stringify whole node. Introduce
get_node_string_from_xpath() which does exactly that.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>