libvirt_domain_disk_remove: Fix disk detach
Thing is that we need to pass full device XML as mandated by
virDomainDetachDeviceFlags() API. So far we are not doing that.
We construct a minimalistic <disk/> XML and hope for the best.
Use freshly introduce get_node_string_from_xpath() to get disk
XML from the domain XML.
Also, fix a copy/paste error in the error message that is
produced when detach fails.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>