libvirt-php: Simplify querying for emulator path
There's this function called connection_get_emulator() which is
supposed to return emulator path for given architecture (or for
the host architecture if none given). It does this by querying
capabilities. Twice. The first time it looks at some bogus XPath
(that never existed), and if that fails, it tries to look at the
correct XPath location. Remove the first attempt which will never
succeed and only burns CPU cycles.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>