libvirt-domain: Actually use array for IP addresses in libvirt_domain_interface_addresses()
If an interface has more than one IP address then iface["addrs"]
should be an array of 'structs' (another associative arrays). But
it's not. Currently, we iterate over ifaces[i]->naddrs and
construct the associative array for one instance of IP address.
But instead of appending the array into an array of IP addresses,
we set iface["addrs"] to this array and therefore overwrite it in
each iteration.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>