gitlab: switch most jobs to build against distro libvirt
The php build needs to validate two axis

 - A variety of libvirt versions
 - A variety of python versions

We get coverage for both these axis by running a build against the
distro provided libvirt packages. All that is then missing is a build
against the latest libvirt git master, which only needs to be run on
a single distro, for which Fedora 32 is picked. CentOS 8 is unable to
build since it lacks php-imagick.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>