Fix signed longlong convert to string as unsigned longlong
The fields in struct _virDomainBlockStats and struct _virDomainInterfaceStats are signed longlong.
Hypervisors may return a field set to -1 which indicates that the hypervisor does not support that statistic.
But libvirt_domain_block_stats() and libvirt_domain_interface_stats() convert these fields to string as unsigned longlong.
This commit fix it.

Signed-off-by: Zhensheng Yuan <extremebored@gmail.com>