libvirt-domain: Switch libvirt_domain_change_memory to virDomainSetMemoryFlags()
The way that libvirt_domain_change_memory() is currently
implemented is: it dumps domain XML, then using couple of
strstr()-s it picks location in the XML where new values for
<memory/> and <currentMemory/> should be placed, replaces the
values and redefines the domain using changed XML.

This is needlessly complicated and error prone. Fortunately,
libvirt already has API that allows changing both memory values:
virDomainSetMemoryFlags().

To keep the function consistent though, we have to return the
domain object again. This made sense when the domain was
redefined (and thus new domain object needed to be returned), but
to keep existing code in mgmt apps working we have to return the
domain object again.

And there's one more caveat: the PHP function accepts @flags as
an optional argument. This was then passed to
virDomainGetXMLDesc(). Obviously, with switch to
virDomainSetMemoryFlags() we could attempt to change config/live
XML, however, that's not what the function did originally. Since
it was redefining domain, it strictly worked over config XML and
thus we don't need this @flags argument anymore (although, it's
previous use is also questionable). Anyway, to keep existing code
working, let's accept the argument but don't use it.

Resolves: https://gitlab.com/libvirt/libvirt-php/-/issues/4
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>