src: Fix arg parsing strings and types
Our APIs accept arguments which are parsed via
zend_parse_parameters() (or macros that wrap around that
function). However, in some cases our format string for arguments
is incorrect (e.g. it contains two or more '|' separators), or
variable where the argument is stored does not correspond to the
type in the formatting string (size_t vs. int).

There's one specific case where this causes a crash
(libvirt_list_networks()) because we are telling the parser
function to store a number in long but pass an int.

These places were identified with a help of
scripts/dev/check_parameters.php script provided by PHP project
(where I had to tweak the regex in check_function() so that it
matches our VIR_XXX_FROM_ARGS() macro too.

After this commit, the script still complains about size_t vs
strsize_t but we should ignore these for the time being, because
of <PHP-7.0.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>