src: Pass connection to resource_change_counter() in destructor
Strictly speaking this is not needed. What's happening is in the
some destructors (e.g. php_libvirt_network_dtor()) when an object
is freed we also let PHP know to decrease the PHP resource
refcounter (by calling resource_change_counter()). Here, as the
second (virConnectPtr) argument NULL is passed. This works for
destructor, because we don't need to do anything with the
connection object. But it goes against the pattern in the rest of
destructors (php_libvirt_domain_dtor(),
php_libvirt_nodedev_dtor(), php_libvirt_nwfilter_dtor(), ...).

Note, after this commit there's still free_resource() which
passes NULL, but that's because we only get a pointer to libvirt
object and not PHP resource, thus we can't get corresponding
connection object. While we might use libvirt API to obtain
connection object (e.g. virDomainGetConnect()), these don't exist
for all the cases (e.g. there's no virStreamGetConnect(), nor
virNodeDeviceGetConnect(), ...) and thus we can stick with NULL
for the time being.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>