Drop support for building on Windows
Support for building on Windows was introduced in 2014 with
commit c49b24ac7241 and never touched afterwards.

The "solution file" (build recipe) seems to have been generated
by Visual C++ 2005 and the libvirt headers are for version 1.2.1,
released in 2013. So even the most recent of the software
components involved are coming up on being ten years old, and
it's very unlikely that they are still useful to someone looking
to build libvirt-php on Windows today.

Moreover, the current implementation relies on binary artifacts
that take up half the size of a release archive in addition to
being questionable from a license compliance standpoint.

Note that this commit only drops the support files and build
system integration: preprocessor conditionals remains sprinkled
all over the place, which makes the change less risky and might
prove helpful to someone trying to revive Windows build support
in the future.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>