Drop LONGLONG_INIT macro
When appending values to an associative array we have macros that
handle long long type of array members: depending on
configuration we may add them as a string or a long. When adding
them as a string we sprintf() the long long value into a buffer
and then append that into the array. And for declaring this
intermediary buffer we have LONGLONG_INIT macro. But this is
needless, because each macro can declare the variable itself.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>