examples: Set filename for screenshots
So far, when a screenshot is downloaded its filename is going to
be random, because we did not set any preferred filename in the
HTTP header. Well, we can do better: not only we can set the
filename to start with the domain name, but we can also provide
the file extension so that it matches the screenshot MIME type.

So far, only PNG and PPM types are supported, because that's what
virsh recognizes too. But we can always extend the list.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>