libvirt-connection: Adapt libvirt_connect_get_soundhw_models() to new QEMU
The libvirt_connect_get_soundhw_models() fetches list of
supported sound devices. It does so by running 'qemu -soundhw
help'. But, starting from QEMU-7.1.0 the '-soundhw' argument is
no longer supported in favor of '-device help'. We need to adopt
to that.

Now, the format in which QEMU reports supported sound devices is
a wee bit less strict than original -soundhw. See for yourself
(qdev_print_devinfo() in QEMU's codebase). But, the first
argument is always the device name, and the description is always
last.

This also fixes the case when the function was called with
CONNECT_FLAG_SOUNDHW_GET_NAMES flag. Previously, this flag was
effectively ignored, because the code looked like this:

  if (inFunc) {
      int i = 0;
      ...

      if ((i > 0) && (flags & CONNECT_FLAG_SOUNDHW_GET_NAMES)) {
          ,,,
      } else {
          ...
      }
  }

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>