src: Introduce typedparams.c
So far we have VIR_TYPED_PARAMETER_ASSOC() which handles
translation for a virTypedParameter to a PHP associative array.
But we are lacking the inverse function - translating an
associative array into set of typed parameters. The reason is
simple - we don't implement a wrapper around any function that
would accept typed parameters. But this will soon change.

The idea is that in PHP the function is called like this:

  libvirt_function(array("param1" => 123, "param2" => true));

which then should create the same set of typed parameters as:

  virTypedParamsAddInt(&params, &nparams, &max, "param1", 123);
  virTypedParamsAddString(&params, &nparams, &max, "param2, true);

Now, it's impossible to know what type to assign to "param1"
value. Therefore, mimic what we do for python bindings and pass
an array of [key, type] pairs to the parser function so that it
knows what type is associated with given key.

From PHP bindings POV this is then used as:

  zval *params = NULL;
  virTypedParameterPtr params2 = NULL;
  int nparams2 = 0;
  virPHPTypedParamsHint hints[] = {
      { "param1", VIR_TYPED_PARAM_INT },
      { "param2", VIR_TYPED_PARAM_BOOLEAN },
  };

  GET_DOMAIN_FROM_ARGS("ra!", &zdomain, &params);

  if (parseTypedParameter(params, &params2, &nparams2, hints, ARRAY_CARDINALITY(hints)) < 0)
      RETURN_FALSE;

  virDomainXXX(domain->domain, params2, nparams2);
  virTypedParamsFree(params2, nparams2);

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>