tests: Make example-qcow2-disk.xml be of type 'qemu'
The example-qcow2-disk.xml defines "test-guest-qcow" domain and
is used only from tests/test-domain-snapshot.phpt. But nothing in
the test requires the domain to be of type 'kvm'. In fact, this
hurts, because if there's a session daemon running and it doesn't
have access to /dev/kvm it's unable to create the domain and the
test fails.

It's perfectly okay when the test domain is of type qemu.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>