libvirt-connection: Rework libvirt_connect_get_maxvcpus()
At PHP level we have libvirt_connect_get_maxvcpus() which is
supposed to be nothing but a thin wrapper over
virConnectGetMaxVcpus(). But it tried to be more. The C function
takes two arguments: virConnect and a domain type (string). While
the PHP wrapper takes only one (the connect resource) and guesses
the second by calling virConnectGetType(). This is obviously
wrong as connection type (libvirt driver name effectually) is
distinct from domain type. For instance, "QEMU" is reported as
connection type, but the QEMU driver accepts "kvm" and "qemu"
domain types.

Ideally, the function would grow $type argument. But that would
break API (though, one can argue this function never worked and
thus is unused).

Nevertheless, the argument can be introduced - as an optional
one.

Resolves: https://gitlab.com/libvirt/libvirt-php/-/issues/9
Signed-off-by: Michal Privoznik <mprivozn@redhat.com>